refactor: synced app's routing with new 'recruit' route segment
