+++ apps/dev-recruiters/tsconfig.json
@@ -30,7 +30,7 @@
     "**/*.tsx",
     "src/components/common/Header/index.js",
     "src/pages/_app.jsx",
-    "src/pages/recruitment.tsx"
+    "src/pages/recruit.tsx"
   ],
   "exclude": ["node_modules"]
 }
